requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-gigachat>=0.2.0
qdrant-client>=1.17.0
sentence-transformers==2.5.1
tiktoken==0.6.0
python-dotenv==1.0.1
pypdf==4.1.0
docx2txt==0.8
requests~=2.32.5
urllib3~=2.6.3